#5068: add require active root
